backend_python_backup/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
eventlet
gunicorn
numpy
scipy
trimesh
opencv-python-headless
scikit-image
Pillow
requests
python-dotenv
stripe